Changelog: 1. Added success msg to reordering in manageNewsBanners.aspx 2. Fixed bug on Dashboard.aspx to display double variables as hours instead of just int 3. Fixed bug on viewModuleInfo.aspx to fix %20 as white spaces in url for deleteBoth.aspx and deleteMaterial.aspx to delete files with %20 in file name. 4. Added fix on downloadTrainingCalendar.aspx.cs to work on safari. Please try if it works.

+++ app/LearnHub/LearnHub/Dashboard.aspx.cs
@@ -376,11 +376,11 @@ namespace LearnHub
             return toReturn;
         }
 
-        public int GetTotalNoOfLearningHours()
+        public double GetTotalNoOfLearningHours()
         {
             string connectionString = ConfigurationManager.ConnectionStrings["DBConnectionString"].ConnectionString;
             DataTable dt = new DataTable();
-            int toReturn = 0;
+            double toReturn = 0.0;
 
             using (SqlConnection connection = new SqlConnection(connectionString))
             {
@@ -397,7 +397,7 @@ namespace LearnHub
                     connection.Close();
                 }
             }
-            toReturn = int.Parse(dt.Rows[0][0].ToString());
+            toReturn = Double.Parse(dt.Rows[0][0].ToString());
             return toReturn;
         }
     }
